(refactor) Completely refactor whole project structure, implement all routes in react router, implement single page app, remove server side rendering except index.html
