+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets" : ["es2015", "react"]
+  "presets": ["react", "es2015", "stage-1"]
 }
