+++ .gitignore
@@ -32,3 +32,11 @@ bower_components/
 *.log
 .DS_Store
 node_modules
+
+/node_modules
+bundle.js
+npm-debug.log
+
+# IntelliJ
+*.iml
+/.idea
