--- hornblowersappfrontend copy/.babelrc
@@ -1,3 +1,0 @@
-{
-  "presets": ["react", "es2015", "stage-1"]
-}
