--- hornblowersappfrontend copy/.gitignore
@@ -1,7 +1,0 @@
-/node_modules
-bundle.js
-npm-debug.log
-
-# IntelliJ
-*.iml
-/.idea
