+++ package.json
@@ -1,31 +1,47 @@
 {
-  "name": "hornblowers",
+  "name": "redux-simple-starter",
   "version": "1.0.0",
-  "description": "",
+  "description": "Simple starter package for Redux with React and Babel support",
   "main": "index.js",
+  "repository": "git@github.com:StephenGrider/ReduxSimpleStarter.git",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node ./node_modules/webpack-dev-server/bin/webpack-dev-server.js",
+    "test": "mocha --compilers js:babel-core/register --require ./test/test_helper.js --recursive ./test",
+    "test:watch": "npm run test -- --watch",
     "dev": "webpack -d --watch",
     "build": "webpack -p",
-    "start": "node src/server/server.js"
+    "start2": "node src/server/server.js"
   },
   "author": "",
   "license": "ISC",
-  "dependencies": {
-    "axios": "^0.15.3",
-    "babel-core": "^6.22.1",
-    "babel-loader": "^6.2.10",
-    "babel-preset-es2015": "^6.22.0",
-    "babel-preset-react": "^6.22.0",
-    "express": "^4.14.1",
-    "mysql": "^2.13.0",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "sequelize": "^3.30.1",
-    "webpack": "^2.2.1",
-    "yelp-fusion": "^1.0.3"
-  },
   "devDependencies": {
+    "axios": "^0.15.3",
+    "babel-core": "^6.2.1",
+    "babel-loader": "^6.2.0",
+    "babel-preset-es2015": "^6.1.18",
+    "babel-preset-react": "^6.1.18",
+    "chai": "^3.5.0",
+    "chai-jquery": "^2.0.0",
+    "jquery": "^2.2.1",
+    "jsdom": "^8.1.0",
+    "mocha": "^2.4.5",
+    "react-addons-test-utils": "^0.14.7",
+    "webpack": "^1.12.9",
+    "webpack-dev-server": "^1.14.0",
     "nodemon": "^1.11.0"
+  },
+  "dependencies": {
+    "babel-preset-stage-1": "^6.1.18",
+    "express": "^4.14.1",
+    "lodash": "^3.10.1",
+    "mysql": "^2.13.0",
+    "react": "^0.14.3",
+    "react-dom": "^0.14.3",
+    "react-redux": "^4.0.0",
+    "react-router": "^3.0.2",
+    "redux": "^3.0.4",
+    "sequelize": "^3.30.1",
+    "yelp-fusion": "^1.0.3",
+    "youtube-api-search": "0.0.5"
   }
 }
