+++ src/client/app/Friend_list_item.js
