+++ src/client/app/YelpResult.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import ReactDOM from 'react-dom';
 
 class YelpResult extends React.Component {
   constructor(props) {
