+++ src/client/app/createNewPage.js
