+++ src/client/app/friend_list.js
