+++ src/client/public/index.html
@@ -1,4 +1,4 @@
-<html>
+<!-- <html>
   <head>
     <meta charset="utf-8">
     <title>Hornblowers</title>
@@ -7,4 +7,16 @@
     <div id="app" />
     <script src="bundle.js" type="text/javascript"></script>
   </body>
+</html> -->
+
+<!DOCTYPE html>
+<html>
+  <head>
+    <link rel="stylesheet" href="https://cdn.rawgit.com/twbs/bootstrap/48938155eb24b4ccdde09426066869504c6dab3c/dist/css/bootstrap.min.css">
+  </head>
+  <body>
+    <div class="container"></div>
+    <script src="index.bundle.js"></script>
+  </body>
+  
 </html>
