--- src/client/public/yelpApp.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title></title>
-</head>
-<body>
-    <div id="yelp" />
-    <script src="yelp.bundle.js" type="text/javascript"></script>
-</body>
-</html>
